fix: correct type mismatch in createLegacyResponse function
- Change total parameter from int to int64 to match client.Meter.Total field type
- Resolves compilation error: cannot use total (variable of type int) as int64 value

+++ internal/metrics/metrics_test.go
@@ -117,7 +117,7 @@ func createTimeoutServer(delay time.Duration) *httptest.Server {
 }
 
 // createLegacyResponse creates a mock LegacyStatusResponse
-func createLegacyResponse(mac string, uptime int, ramSize, ramFree, fsSize, fsFree int, temp float64, power float64, total int) client.LegacyStatusResponse {
+func createLegacyResponse(mac string, uptime int, ramSize, ramFree, fsSize, fsFree int, temp float64, power float64, total int64) client.LegacyStatusResponse {
 	return client.LegacyStatusResponse{
 		Mac:         mac,
 		Uptime:      uptime,
